refactored pastel colors for strings, moved videos modules
