+++ scripts/node/providers/youtube/cli/lib/videos/videos-validate-yargs.js
